Break claims out in to own package
Move them in to their own files too
